Add status revised to view

+++ app/views/TraineeView.php
@@ -25,7 +25,9 @@
                                 && $report->status() !== Report::STATUS_APPROVED
                                 && $report->status() !== Report::STATUS_APPROVAL_REQUESTED): ?>
                             <li><a href="/report/requestApproval?reportId=<?php echo $reportId; ?>&action=requestApproval" onclick="return confirm('Soll der Bericht eingereicht werden?')">Einreichen</a></li>
-                            <li><a href="/report/delete?reportId=<?php echo $reportId; ?>&action=delete" onclick="return confirm('Soll der Bericht wirklich gelöscht werden?')">Löschen</a></li>
+                            <?php if ($report->status() !== Report::STATUS_REVISED): ?>
+                                <li><a href="/report/delete?reportId=<?php echo $reportId; ?>&action=delete" onclick="return confirm('Soll der Bericht wirklich gelöscht werden?')">Löschen</a></li>
+                            <?php endif ?>
                         <?php endif; ?>
                         <?php if ($report->status() === Report::STATUS_APPROVED
                                 || $report->status() === Report::STATUS_APPROVAL_REQUESTED): ?>
